update the forked repository

+++ src/rcb.h
@@ -14,7 +14,6 @@
 #ifndef LAMMPS_RCB_H
 #define LAMMPS_RCB_H
 
-#include <mpi.h>
 #include "pointers.h"
 
 namespace LAMMPS_NS {
@@ -44,7 +43,7 @@ class RCB : protected Pointers {
   void compute(int, int, double **, double *, double *, double *);
   void compute_old(int, int, double **, double *, double *, double *);
   void invert(int sortflag = 0);
-  bigint memory_usage();
+  double memory_usage();
 
   // DEBUG methods
   //void check();
